Leetcode Contest Ranking Page (#60)
* Codeforces api working

* PostgreSQL Supabase Database

* Some changes

* Added leetcode rankings page

* PostgreSQL Supabase Database

* Leetcode institute ranking added

* trying vercel deployment

* trying railway for deployment

* Added some requirements

* Added some requirements

* Update requirements.txt

* Create requirements.txt

* Changes for deployment

* Delete requirements.txt

* Update requirements.txt

* Update settings.py

* Update views.py

* logger info in views

* Added a new leetcode-contest-ranking page

* Improved computational speed of leetcoderanking url

* Contest details in db

* CCPS Leetcode rankings

* CCPS logo added in navbar

* UI changes

* added feature of leetcode contest ranking

---------

+++ api/leaderboard/models.py
@@ -1,6 +1,10 @@
 from django.db import models
 from datetime import datetime, timezone, timedelta
 from django.contrib.auth.models import User
+from datetime import datetime
+
+from django.contrib.postgres.fields import ArrayField
+
 
 class githubUser(models.Model):
     username = models.CharField(max_length=64, unique=True)
@@ -47,7 +51,7 @@ class codeforcesUser(models.Model):
     username = models.CharField(max_length=64, unique=True)
     max_rating = models.PositiveIntegerField(default=0)
     rating = models.PositiveIntegerField(default=0)
-    last_activity = models.PositiveIntegerField(
+    last_activity = models.BigIntegerField(
         default=datetime.max.timestamp()
     )
     last_updated = models.DateTimeField(auto_now=True)
@@ -107,7 +111,7 @@ class codeforcesUserRatingUpdate(models.Model):
     index = models.PositiveIntegerField(default=0)
     prev_index = models.PositiveIntegerField(default=0)
     rating = models.PositiveIntegerField(default=0)
-    timestamp = models.PositiveIntegerField(default=0)
+    timestamp = models.BigIntegerField(default=0)
     def __str__(self):
         return f"{self.cf_user.username}.{self.index} {self.rating}"
 
@@ -158,3 +162,4 @@ class CodechefFriends(models.Model):
 class OpenlakeFriends(models.Model):
     user =models.ForeignKey(User,on_delete=models.CASCADE,null=True)
     olFriend_uname=models.CharField(max_length=64)
+
